Minor code prettyfication
