+++ rabbus.go
@@ -33,9 +33,6 @@ const (
 )
 
 type (
-	// Option represents an option you can pass to New.
-	// See the documentation for the individual options.
-	Option func(*Rabbus) error
 	// OnStateChangeFunc is the callback function when circuit breaker state changes.
 	OnStateChangeFunc func(name, from, to string)
 
@@ -226,7 +223,7 @@ func (r *Rabbus) Run(ctx context.Context) error {
 				return nil
 			}
 
-			r.handleAmqpClose(err)
+			r.handleAMQPClose(err)
 
 			// We have reconnected, so we need a new NotifyClose again.
 			notifyClose = r.NotifyClose(make(chan *amqp.Error))
@@ -274,7 +271,7 @@ func (r *Rabbus) Listen(c ListenConfig) (chan ConsumerMessage, error) {
 
 	messages := make(chan ConsumerMessage, 256)
 	go r.wrapMessage(c, msgs, messages)
-	go r.listenReconn(c, messages)
+	go r.listenReconnect(c, messages)
 
 	return messages, nil
 }
@@ -331,128 +328,13 @@ func (r *Rabbus) produce(m Message) {
 	r.emitOk <- struct{}{}
 }
 
-// Durable indicates of the queue will survive broker restarts. Default to true.
-func Durable(durable bool) Option {
-	return func(r *Rabbus) error {
-		r.config.durable = durable
-		return nil
-	}
-}
-
-// PassiveExchange forces passive connection with all exchanges using
-// amqp's ExchangeDeclarePassive instead the default ExchangeDeclare
-func PassiveExchange(isExchangePassive bool) Option {
-	return func(r *Rabbus) error {
-		r.config.isExchangePassive = isExchangePassive
-		return nil
-	}
-}
-
-// PrefetchCount limit the number of unacknowledged messages.
-func PrefetchCount(count int) Option {
-	return func(r *Rabbus) error {
-		r.config.qos.prefetchCount = count
-		return nil
-	}
-}
-
-// PrefetchSize when greater than zero, the server will try to keep at least
-// that many bytes of deliveries flushed to the network before receiving
-// acknowledgments from the consumers.
-func PrefetchSize(size int) Option {
-	return func(r *Rabbus) error {
-		r.config.qos.prefetchSize = size
-		return nil
-	}
-}
-
-// QosGlobal when global is true, these Qos settings apply to all existing and future
-// consumers on all channels on the same connection. When false, the Channel.Qos
-// settings will apply to all existing and future consumers on this channel.
-// RabbitMQ does not implement the global flag.
-func QosGlobal(global bool) Option {
-	return func(r *Rabbus) error {
-		r.config.qos.global = global
-		return nil
-	}
-}
-
-// Attempts is the max number of retries on broker outages.
-func Attempts(attempts int) Option {
-	return func(r *Rabbus) error {
-		r.config.retryCfg.attempts = attempts
-		return nil
-	}
-}
-
-// Sleep is the sleep time of the retry mechanism.
-func Sleep(sleep time.Duration) Option {
-	return func(r *Rabbus) error {
-		if sleep == 0 {
-			r.config.retryCfg.reconnectSleep = time.Second * 10
-		}
-		r.config.retryCfg.sleep = sleep
-		return nil
-	}
-}
-
-// BreakerInterval is the cyclic period of the closed state for CircuitBreaker to clear the internal counts,
-// If Interval is 0, CircuitBreaker doesn't clear the internal counts during the closed state.
-func BreakerInterval(interval time.Duration) Option {
-	return func(r *Rabbus) error {
-		r.config.breaker.interval = interval
-		return nil
-	}
-}
-
-// BreakerTimeout is the period of the open state, after which the state of CircuitBreaker becomes half-open.
-// If Timeout is 0, the timeout value of CircuitBreaker is set to 60 seconds.
-func BreakerTimeout(timeout time.Duration) Option {
-	return func(r *Rabbus) error {
-		r.config.breaker.timeout = timeout
-		return nil
-	}
-}
-
-// Threshold when a threshold of failures has been reached, future calls to the broker will not run.
-// During this state, the circuit breaker will periodically allow the calls to run and, if it is successful,
-// will start running the function again. Default value is 5.
-func Threshold(threshold uint32) Option {
-	return func(r *Rabbus) error {
-		if threshold == 0 {
-			threshold = 5
-		}
-		r.config.breaker.threshold = threshold
-		return nil
-	}
-}
-
-// OnStateChange is called whenever the state of CircuitBreaker changes.
-func OnStateChange(fn OnStateChangeFunc) Option {
-	return func(r *Rabbus) error {
-		r.config.breaker.onStateChange = fn
-		return nil
-	}
-}
-
-// AmqpProvider expose a interface for interacting with amqp broker
-func AmqpProvider(provider AMQP) Option {
-	return func(r *Rabbus) error {
-		if provider != nil {
-			r.AMQP = provider
-			return nil
-		}
-		return errors.New("unexpected amqp provider")
-	}
-}
-
 func (r *Rabbus) wrapMessage(c ListenConfig, sourceChan <-chan amqp.Delivery, targetChan chan ConsumerMessage) {
 	for m := range sourceChan {
 		targetChan <- newConsumerMessage(m)
 	}
 }
 
-func (r *Rabbus) handleAmqpClose(err error) {
+func (r *Rabbus) handleAMQPClose(err error) {
 	for {
 		time.Sleep(time.Second)
 		aw, err := amqpWrap.New(r.config.dsn, r.config.isExchangePassive)
@@ -480,7 +362,7 @@ func (r *Rabbus) handleAmqpClose(err error) {
 	}
 }
 
-func (r *Rabbus) listenReconn(c ListenConfig, messages chan ConsumerMessage) {
+func (r *Rabbus) listenReconnect(c ListenConfig, messages chan ConsumerMessage) {
 	for range r.reconn {
 		msgs, err := r.CreateConsumer(c.Exchange, c.Key, c.Kind, c.Queue, r.config.durable, c.DeclareArgs.args, c.BindArgs.args)
 		if err != nil {
@@ -488,7 +370,7 @@ func (r *Rabbus) listenReconn(c ListenConfig, messages chan ConsumerMessage) {
 		}
 
 		go r.wrapMessage(c, msgs, messages)
-		go r.listenReconn(c, messages)
+		go r.listenReconnect(c, messages)
 		break
 	}
 }
@@ -501,6 +383,7 @@ func newBreakerSettings(c config) gobreaker.Settings {
 	s.ReadyToTrip = func(counts gobreaker.Counts) bool {
 		return counts.ConsecutiveFailures > c.breaker.threshold
 	}
+
 	if c.breaker.onStateChange != nil {
 		s.OnStateChange = func(name string, from gobreaker.State, to gobreaker.State) {
 			c.breaker.onStateChange(name, from.String(), to.String())
